break apart the cucumber runner so it can be used elsewhere"
