+++ pkg/reconciler/testing/featured/relative.go
@@ -1,0 +1,35 @@
+/*
+Copyright 2021 The Knative Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package featured
+
+import (
+	"path/filepath"
+	"runtime"
+	"strings"
+)
+
+var prefixPath string
+
+func init() {
+	_, fileName, _, _ := runtime.Caller(0)
+	prefixPath = filepath.ToSlash(filepath.Dir(filepath.Dir(fileName))) + "/"
+}
+
+func relative() string {
+	_, fn, _, _ := runtime.Caller(1)
+	return strings.TrimPrefix(filepath.ToSlash(fn), prefixPath)
+}
